webrtc/requirements.txt
aiohttp
aiortc
python-socketio[client]
